Teeny-tiny-llama/requirements.txt
transformers
torch
pyyaml
datasets
wandb
codecarbon
huggingface_hub
accelerate